ETU-17005: Starting to put common code from jwt-client-spring and jwt-client-spring-cloud in abstract classes in jwt-client-spring-core
