+++ jwt-client/jwt-client-spring-cloud/src/main/java/org/entur/jwt/client/properties/JwtClientProperties.java
@@ -1,0 +1,6 @@
+package org.entur.jwt.client.properties;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class JwtClientProperties extends AbstractSpringJwtClientProperties {}
