+++ jwt-client/jwt-client-spring-cloud/src/main/java/org/entur/jwt/client/springcloud/EagerContextStartedListener.java
@@ -1,0 +1,19 @@
+package org.entur.jwt.client.springcloud;
+
+import org.entur.jwt.client.AccessTokenProvider;
+import org.entur.jwt.client.springcore.AbstractEagerContextStartedListener;
+
+import java.util.Map;
+
+/**
+ *
+ * This kicks off the eager JWT caching, if enabled. If fetching JWTs fail, a warning is logged.
+ *
+ */
+
+public class EagerContextStartedListener extends AbstractEagerContextStartedListener {
+
+    public EagerContextStartedListener(Map<String, AccessTokenProvider> providersById) {
+        super(providersById);
+    }
+}
