+++ jwt-client/jwt-client-spring-cloud/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.entur.jwt.client.springcloud.JwtClientAutoConfiguration
