+++ jwt-client/jwt-client-spring-cloud/src/test/resources/application-auth0-multiple.properties
