+++ jwt-client/jwt-client-spring-cloud/src/test/resources/application-keycloak.properties
