+++ jwt-client/jwt-client-spring-cloud/src/test/resources/application.properties
