+++ jwt-client/jwt-client-spring-cloud/src/test/resources/auth0ClientCredentialsResponse2.json
