+++ jwt-client/jwt-client-spring-cloud/src/test/resources/keycloakClientCredentialsResponse.json
