+++ jwt-client/jwt-client-spring-core/README.md
@@ -1,0 +1,23 @@
+# jwt-client-spring
+JWT spring client core implementation.
+
+See [Auth0SingleClientTest] for a mocking example.
+
+## Maven / Gradle coordinates
+Maven coordinates:
+
+```xml
+<dependency>
+    <groupId>org.entur.jwt-rs</groupId>
+    <artifactId>jwt-client-spring-cloud</artifactId>
+    <version>${jwt-rs.version}</version>
+</dependency>
+```
+
+Gradle coordinates:
+
+```groovy
+api("org.entur.jwt-rs:jwt-client-spring:${jwtResourceServerVersion}")
+```
+
+[Auth0SingleClientTest]: src/test/java/org/entur/jwt/client/springreactive/Auth0SingleClientTest.java
