--- jwt-client/jwt-client-spring-reactive/src/main/java/org/entur/jwt/client/springreactive/SpringJwtClientProperties.java
@@ -1,15 +1,0 @@
-package org.entur.jwt.client.springreactive;
-
-import org.entur.jwt.client.properties.JwtClientProperties;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- *
- * Extend properties to get the tooling working for Spring properties.
- *
- */
-
-@ConfigurationProperties(prefix = "entur.jwt.clients")
-public class SpringJwtClientProperties extends JwtClientProperties {
-
-}
