--- jwt-client/jwt-client-spring-reactive/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.entur.jwt.client.springreactive.JwtClientAutoConfiguration
