--- jwt-client/jwt-client-spring-reactive/src/test/java/org/entur/jwt/client/springreactive/Auth0MultipleClientsTest.java
@@ -1,107 +1,0 @@
-package org.entur.jwt.client.springreactive;
-
-import okhttp3.mockwebserver.MockResponse;
-import okhttp3.mockwebserver.MockWebServer;
-import org.entur.jwt.client.AccessToken;
-import org.entur.jwt.client.AccessTokenProvider;
-import org.entur.jwt.client.springreactive.actuate.AccessTokenProviderHealthIndicator;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.core.io.Resource;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.annotation.DirtiesContext.ClassMode;
-import org.springframework.test.context.TestPropertySource;
-
-import java.io.IOException;
-
-import static com.google.common.truth.Truth.assertThat;
-import static io.restassured.RestAssured.given;
-import static org.entur.jwt.client.springreactive.TestUtils.asString;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@TestPropertySource(locations = "/application-auth0-multiple.properties")
-@DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
-public class Auth0MultipleClientsTest {
-
-    @LocalServerPort
-    private int randomServerPort;
-
-    @Autowired
-    @Qualifier("firstClient")
-    private AccessTokenProvider firstAccessTokenProvider;
-
-    @Autowired
-    @Qualifier("secondClient")
-    private AccessTokenProvider secondAccessTokenProvider;
-
-    @Autowired
-    private AccessTokenProviderHealthIndicator healthIndicator;
-
-    @Value("classpath:auth0ClientCredentialsResponse1.json")
-    private Resource resource1;
-
-    @Value("classpath:auth0ClientCredentialsResponse2.json")
-    private Resource resource2;
-
-    private MockWebServer mockWebServer;
-
-    @BeforeEach
-    void setUp() throws IOException {
-        mockWebServer = new MockWebServer();
-        mockWebServer.start(8000);
-        mockWebServer.url("/oauth/token");
-    }
-
-    @AfterEach
-    void tearDown() throws IOException {
-        mockWebServer.shutdown();
-    }
-
-    @Test
-    public void contextLoads() {
-        assertNotNull(firstAccessTokenProvider);
-        assertNotNull(secondAccessTokenProvider);
-        assertNotNull(healthIndicator);
-    }
-
-    @Test
-    public void testAccessToken() throws Exception {
-
-        mockWebServer.enqueue(new MockResponse().setBody(asString(resource1)));
-        mockWebServer.enqueue(new MockResponse().setBody(asString(resource2)));
-
-        AccessToken accessToken1 = secondAccessTokenProvider.getAccessToken(false);
-
-        assertThat(accessToken1.getType()).isEqualTo("Bearer");
-        assertThat(accessToken1.getValue()).isEqualTo("a.b.c");
-
-        AccessToken accessToken2 = firstAccessTokenProvider.getAccessToken(false);
-
-        assertThat(accessToken2.getType()).isEqualTo("Bearer");
-        assertThat(accessToken2.getValue()).isEqualTo("d.e.f");
-    }
-
-    @Test
-    public void testActuator() throws Exception {
-        // down
-        mockWebServer.enqueue(new MockResponse().setResponseCode(HttpStatus.NOT_FOUND.value()));
-        mockWebServer.enqueue(new MockResponse().setResponseCode(HttpStatus.NOT_FOUND.value()));
-
-        // up
-        mockWebServer.enqueue(new MockResponse().setBody(asString(resource1)));
-        mockWebServer.enqueue(new MockResponse().setBody(asString(resource2)));
-
-        given().port(randomServerPort).log().all().when().get("/actuator/health/readiness").then().log().all().assertThat().statusCode(HttpStatus.SERVICE_UNAVAILABLE.value());
-        given().port(randomServerPort).log().all().when().get("/actuator/health/readiness").then().log().all().assertThat().statusCode(HttpStatus.OK.value());
-
-    }
-}
