--- jwt-client/jwt-client-spring-reactive/src/test/java/org/entur/jwt/client/springreactive/TestApplication.java
@@ -1,11 +1,0 @@
-package org.entur.jwt.client.springreactive;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(TestApplication.class, args).start();
-    }
-}
