--- jwt-client/jwt-client-spring-reactive/src/test/java/org/entur/jwt/client/springreactive/TestUtils.java
@@ -1,22 +1,0 @@
-package org.entur.jwt.client.springreactive;
-
-import org.springframework.core.io.Resource;
-import org.springframework.util.FileCopyUtils;
-
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.io.UncheckedIOException;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-public class TestUtils {
-
-    public static String asString(Resource resource) {
-        try (Reader reader = new InputStreamReader(resource.getInputStream(), UTF_8)) {
-            return FileCopyUtils.copyToString(reader);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-}
